build: validate Linux build and add CI job
- Fix implicit-declaration warnings in process.c: add missing
  #include "prep.h" and #include "exec.h"
- Fix type mismatch: mp3gain_prepare_first_mp3_frame frame param
  changed from int* to unsigned long* in process.c and process.h
- Remove unused variables flags/curFieldNum in apetag.c and
  freqidx in mp3gain.c changeGain (set but never read)
- Make smoke_test.sh cross-platform: auto-detect binary path for
  Linux (build/mp3gain) and Windows (build/Release/mp3gain.exe)
- Add build-linux CI job (ubuntu-latest, gcc, libmpg123-dev)
  with configure, build, smoke tests and artifact upload
- Update PROJECT.md and HANDOFF.md with Linux build status

Linux build: zero warnings, 24/24 smoke tests passing.

+++ project/process.c
@@ -4,8 +4,10 @@
 
 #include "apetag.h"
 #include "cli.h"
+#include "exec.h"
 #include "gain_analysis.h"
 #include "mp3gain.h"
+#include "prep.h"
 #include <mpg123.h>
 
 #include "process.h"
@@ -363,7 +365,7 @@ unsigned long mp3gain_prepare_first_mp3_frame(
     int *mode,
     int *mpegver,
     int *freqidx,
-    int *frame,
+    unsigned long *frame,
     long arrbytesinframe[16]
 )
 {
